+++ crates/leimu-mem/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Cangelog
 All changes to this project will be documented in this file.
+## [0.2.3] - 2026-04-26
+
+### Added
+- insert_with_index and try_insert_with_index for SlotMap
+- assume_init_array for ArrayVec
 ## [0.2.2] - 2026-04-25
 ### Added
 
@@ -22,6 +27,7 @@ All changes to this project will be documented in this file.
 - Custom containers
 - Arena allocator implementation
 
+[0.2.3]: https://crates.io/crates/leimu-mem/0.2.3
 [0.2.2]: https://crates.io/crates/leimu-mem/0.2.2
 [0.2.1]: https://crates.io/crates/leimu-mem/0.2.1
 [0.2.0]: https://crates.io/crates/leimu-mem/0.2.0